Moved Generics to instance system
GenericBases are defined by the IDM code, and contain a GenericType, which acts as a restrictor on what can be applied to the generic and a placeholder in member signatures
GenericInstances are what happens when a GenericBase is populated by a type which matches GenericType. They are a generic structure + an actual type.